Add django-cas-client version 1.5.2 to this repository
Since django-cas-client is not available in standard Debian, here is all the
'cas' code, used for the Single-Sign-On procedure. Previously, this code was
compiled and published as a seperate TiU-only debian package, but I find this
too much hassle for a few Python files. Note that django-cas-client's MIT
license permits this inclusion.

For completeness, here is the exact commit that this code was copied from:
https://github.com/kstateome/django-cas/tree/1.5.2