Remove django-cleanup dependency
Since this is now removed, files deleted from the database will not be
automatically deleted from disk. This was a dangerous antipattern anyway.
In the future, a good solution would be to create custom management commands
for the following tasks:

- Deleting unused files
- Deleting users that have not logged in for over a year
  (per university's requirements)

A cronjob could then take care of periodically running these commands.
To be continued...