Port code to be compatible for Django 2 and up
The main update for Django 2.x compatibility was adding the required
`on_delete` argument to ForeignKey fields. The historical migrations also
had to be rewritten to make the use of the formerly implicit CASCADE
explicit. In 3 cases I have changed the default cascade behavior and added
a new migration for this.